LauncherActivity
Develop an Android app as per the following specifications:
i.	There should be three activities. First activity is the launcher activity. The second activity must have the image of College in full screen. The Third activity must display the Login form with Username and Password using Frame Layout and Toast valid credentials
ii.	In the launcher activity, there should be a Textbox with the text “About College”
iii.	In the launcher activity there should be two buttons with the labels “Campus” and “Login”.
iv.	When “Campus” button is tapped, the activity with college image must be opened. When “Login” button is tapped, the activity with the Login Form must be opened.